+++ CHANGELOG.md
@@ -5,6 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+## [0.3.0] - 2026-08-01
+### Removed
+- Import of `re`, `json`, and `asyncio` packages
+### Changed
+- Updated `fastmcp` from `>= 2.12.4` to `== 3.4.5`
+- Updated `mcp` from `== 1.16.0` to `== 1.29.0`
+### Fixed
+- Sometimes "No such place found" would cause invalid structured content to be returned by the tool causing an MCP error. Fix was to not for a `json.dumps(...)`
+
 ## [0.2.0] - 2025-10-07
 ### Added
 - `fastmcp.json` to adhere to changes in FastMCP module updates